requirements.txt
selenium==4.15.2
requests==2.31.0
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0

lxml==4.9.3